requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-groq
langgraph
cassio
python-dotenv
pydantic
typing-extensions